Select a set of minimum MAE values for interpolation.
I selected the ~10 smallest values of MAE from the cross-validation for each season and am going to interpolate surfaces using the parameter values for each. This will be my "initial" run of interpolation to see if I get a reasonable pattern with these values. I also aim to demonstrate the influences of each parameter. For the 2D interpolation there is just the tension and smoothing paramter, but for the 3D interpolation there is also the zscale parameter.